pyproject.toml
[project]
name = "slm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "pytest>=8.4.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
